requirements.txt
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contextvars==2.4
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
htmltools==0.2.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
immutables==0.19
importlib-metadata==6.6.0
ipykernel==6.23.0
ipython==8.12.2
ipywidgets==8.0.6
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
linkify-it-py==2.0.2
lxml==4.9.2
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
multitasking==0.0.11
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
python-bcb==0.1.8
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
rfc3986==1.5.0
rsconnect-python==1.10.0
scipy==1.10.1
semver==2.13.0
shiny==0.3.3
shinywidgets==0.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.26.1
statsmodels==0.14.0
tenacity==8.2.2
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.2
uvicorn==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
widgetsnbextension==4.0.7
yfinance==0.2.18
zipp==3.15.0
